+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Attribution text cursor feedback and font update
+    - Attribution text now displays pointer cursor when hovered
+    - Changed font from Saira-Regular to Saira-Light
+
 ## [0.0.22] - 2025-08-04
 
 ## [0.0.21] - 2025-08-04
@@ -289,10 +295,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Platform-specific configuration paths (XDG compliant)
 
 [Unreleased]: https://github.com/emilyst/stardrift/compare/v0.0.22...HEAD
+
 [0.0.22]: https://github.com/emilyst/stardrift/compare/v0.0.21...v0.0.22
+
 [0.0.21]: https://github.com/emilyst/stardrift/compare/v0.0.20...v0.0.21
+
 [0.0.20]: https://github.com/emilyst/stardrift/compare/v0.0.19...v0.0.20
+
 [0.0.19]: https://github.com/emilyst/stardrift/compare/v0.0.18...v0.0.19
+
 [0.0.18]: https://github.com/emilyst/stardrift/compare/v0.0.17...v0.0.18
 
 [0.0.17]: https://github.com/emilyst/stardrift/compare/v0.0.16...v0.0.17